requirements.txt
# Kid Fax SMS Mailbox Dependencies
python-escpos==3.0a8
pyusb==1.2.1
Pillow==10.3.0
twilio==9.0.1
pynput==1.7.6
